+++ lib/widgets/sheets/add_channel_flow.dart
@@ -5,7 +5,7 @@ import 'package:twake/blocs/profile_bloc/profile_bloc.dart';
 import 'package:twake/blocs/user_bloc/user_bloc.dart';
 import 'package:twake/repositories/add_channel_repository.dart';
 import 'package:twake/widgets/sheets/channel_info_form.dart';
-import 'package:twake/widgets/sheets/channel_participants_list.dart';
+import 'package:twake/widgets/sheets/participants_list.dart';
 
 class AddChannelFlow extends StatelessWidget {
   @override
@@ -15,7 +15,7 @@ class AddChannelFlow extends StatelessWidget {
       ChannelInfoForm(),
       BlocProvider<UserBloc>(
         create: (_) => UserBloc(ProfileBloc.userId),
-        child: ChannelParticipantsList(),
+        child: ParticipantsList(),
       ),
     ];
     return BlocBuilder<AddChannelBloc, AddChannelState>(
