+++ lib/widgets/sheets/add_direct_flow.dart
@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:twake/blocs/profile_bloc/profile_bloc.dart';
 import 'package:twake/blocs/user_bloc/user_bloc.dart';
-import 'package:twake/widgets/sheets/channel_participants_list.dart';
+import 'package:twake/widgets/sheets/participants_list.dart';
 
 class AddDirectFlow extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocProvider<UserBloc>(
       create: (_) => UserBloc(ProfileBloc.userId),
-      child: ChannelParticipantsList(isDirect: true),
+      child: ParticipantsList(isDirect: true),
     );
   }
 }
