+++ lib/widgets/sheets/add_workspace_flow.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:twake/widgets/sheets/channel_participants_list.dart';
+import 'package:twake/widgets/sheets/participants_list.dart';
 
 class AddWorkspaceFlow extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return ChannelParticipantsList(isDirect: true);
+    return ParticipantsList(isDirect: true);
   }
 }
