+++ lib/widgets/sheets/draggable_scrollable.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:twake/blocs/add_channel_bloc/add_channel_bloc.dart';
-import 'package:twake/blocs/sheet_bloc/sheet_bloc.dart';
 import 'package:twake/repositories/add_channel_repository.dart';
 import 'package:twake/repositories/sheet_repository.dart';
 import 'package:twake/widgets/sheets/add_channel_flow.dart';
