+++ lib/widgets/sheets/name_container.dart
@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 import 'package:twake/config/dimensions_config.dart';
 import 'package:twake/widgets/common/selectable_avatar.dart';
-import 'package:twake/widgets/sheets/channel_info_text_form.dart';
+import 'package:twake/widgets/sheets/sheet_text_field.dart';
 
-class ChannelNameContainer extends StatelessWidget {
+class NameContainer extends StatelessWidget {
   final TextEditingController controller;
   final FocusNode focusNode;
 
-  const ChannelNameContainer({
+  const NameContainer({
     Key key,
     @required this.controller,
     @required this.focusNode,
@@ -18,30 +18,15 @@ class ChannelNameContainer extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      // height: 83.0,
       width: MediaQuery.of(context).size.width,
       child: Container(
         padding: EdgeInsets.symmetric(horizontal: 12.0),
         color: Colors.white,
-        child: ChannelInfoTextForm(
+        child: SheetTextField(
           hint: 'Channel name',
           controller: controller,
           focusNode: focusNode,
         ),
-        // child: Row(
-        //   children: [
-        //     // Expanded(child: SelectableAvatar()),
-        //     Expanded(
-        //       flex: 3,
-        //       child: ChannelInfoTextForm(
-        //         hint: 'Channel name',
-        //         controller: controller,
-        //         focusNode: focusNode,
-        //       ),
-        //     ),
-        //     SizedBox(width: 10.0),
-        //   ],
-        // ),
       ),
     );
   }
